MXS-3665: List ciphers on cipher list conflicts
If the client offers a set of ciphers that the server doesn't accept,
OpenSSL will report a `no shared cipher` error. As was described in the
initial issue, this tells why it failed but it lacks the required
information to identify and fix the source of the problem.

By listing both the ciphers the client offered (only OpenSSL 1.1.1) and
the ones we accept, the administrator should be able to deduce why the
TLS connection failed.