MXS-3990 A mxb::Worker::Object must know its worker
With this change the delayed call ids can be stored in the
Worker::Object and it can automatically cancel all pending
delayed calls when it is destructed. This means that the
Reference arrangement can be removed.

Furthermore, since Worker::Object will know its delayed calls,
it is possible to add functionality for disabling and enabling
all current delayed calls, something which comes in handly if
a Worker::Object is moved from one worker to another.