MXS-3982: Add TLS certificate reloading endpoint
The /maxscale/tls/reload endpoint will reconfigure all listeners and
servers with their existing configurations. This will cause the
certificates to be reloaded again after they have been validated.

Compared to simply recreationg the SSL_CTX object, this is a
comparatively heavy operation to perform. The benefit of reloading the
configuration is that we leverage the existing validation and
configuration code with only minimal changes required.

Added the MaxCtrl command `reload tls` that will call the new API
endpoint.