MXS-3990 Register dcall ids in owning object
Now an object is aware of all delayed calls that are associated
with it. That makes it possible to automatically cancel all
pending calls when the object is destructed. A properly written
object would do that explicitly, but in case it's not done, it's
now done implicitly. A failure to do that would earlier have
caused a crash, so logging an error in that case should not flood
the log and, if it occurs, provide an explanation for crashes
whose cause has not been figured out.