MXS-3960 Add `handleRevertDiagonal` function props
By default, the diagonal creates an obtuse shape line where it draws
from the source point to the vertical middle point (middle point
between source.y and target.y) as a straight line. Then it draws
from that midpoint to the source point which is perpendicular to
the source node. The `config_graph_data` treats the monitors as
the root nodes so the monitor is a source point while the service
is a target point, the diagonal needs to be reversed in this case.

This adds mentioned props to handle revert the diagonal.