MXS-3413: Send warnings in HTTP responses
If a request done via the REST API generates warnings, it will now be
available to the client application. Right now only MaxCtrl will display
these warnings to the user.

As modifying static objects at runtime is a perfectly valid, the
warnings can be disabled for MaxCtrl by creating the
`~/.maxctrl-no-warnings` file.