MXS-3962: Save only modified values
This should make upgrades easier and make the runtime reconfiguration of
the global [maxscale] section much more convenient for end users.

Adjusted the wording in the warning message to state that only the
modified values will override the static ones.

Modified test_server to use the constructed objects for comparison
instead of the stored config values. This only matters for the port as
it uses the default value which is now no longer saved into the config
file.