MXS-3960 Calculate nodeSize automatically for `dag-graph`
Dynamic height is typically needed. With the current approach
that is to get the height of each node outside of `dag-graph`
and pass it down to `dag-graph` is not reliable. When
a node is deleted, that node should be removed from `dynNodeHeightMap`
but it doesn't. The better way is to get node height in `dag-graph`
after rendering it in the next tick.

This calculates node height in `dag-graph` when `dynNodeHeight` props
is enabled.