MXS-3413: Make volatile servers their own type
This fixes a debug assertion which assumed all names are either in a
static or a dynamic config and volatile servers created by xpandmon are
in neither.

The same debug assertion would also be hit when configuration
synchronization was being used. This happens because the objects are not
stored in the same manner as they are with the traditional usage.

When the configuration synchronization is enabled, it is not easy to
know where a change originated from and thus it should not be confused
with the normal runtime configuration. Labeling the source type as
"cluster" makes this clear.

If load_persisted_configs is disabled, the code does not know whether an
object was modified at runtime or not, only whether it originated from a
static file or not.