MXS-3960 Use `sugiNodeSize` operator to enable dynamic node size layout
The service node will increase its height to shows filters information.
Just in case a service uses dozen of filters, it's better to have a button
to show it on will. At the moment, the graph uses `nodeSize`
operator which doesn't support dynamic nodeSize. As a result, all nodes
are equally packed and left with plenty spaces.

This uses the `sugiNodeSize` operator which supports dynamic node size.