MXS-4046: Use COM_RESET_CONNECTION if possible
If the account and current database of the client connection match a
connection in the pool, the COM_RESET_CONNECTION packet can be used to
quickly reset the connection.

The benefits of doing this could be considered a marginal improvement:
apart from the smaller amount of data and the lack of needing to perform
authentication, the most of the time is still spent waiting for the
server to respond. However, due to MDEV-28059 this ends up being twice
as fast as a COM_CHANGE_USER as it avoids the extra roundtrip that
always occurs with it even if the connection uses the exact same
credentials.

Further improvements to this can be made. Detecting situations where the
connection can be reused without resetting the session state would
reduce the number of roundtrips to zero. A similar approach could be
taken by sending the stored history in order to overwrite the connection
state. Since the history is sent asynchronously in a streaming fashion,
this will effectively eliminate the need for a roundtrip before queries
are accepted.

Returning an abstract integer from the can_reuse method instead of an
enumeration value makes it possible to extend the implementation without
changing the APIs. This also decouples the protocol implementation from
the generic protocol API.