MXS-4046: Send history as soon as possible
If the connection is being reset with a COM_RESET_CONNECTION, the
session command history can be sent without waiting for the server's
response. This cannot be done if a COM_CHANGE_USER is used to reset the
connection as the server might request reauthentication by sending an
AuthSwitchRequest packet.

A micro-optimization could be done for the writing of the packets in
this case: by not flushing the backend DCB's writeq, it would allow both
the COM_RESET_CONNECTION and the history to be sent in the same
drain_write() call if they can fit into the network buffers.