MXS-4045: Add `create report`
The command gets the contents of all the relevant endpoints and saves
the output into a file. This should help speed up problem resolution by
providing a single command with which the whole MaxScale state can be
retrieved.

For version 6, this can also include the log data as long as MaxScale is
configured to write the log entries into a file or syslog.