Add scripts for regenerating REST API documentation
The old python code required a bunch of manual steps to work and a
custom build of MaxScale. By using an existing docker image, the whole
process is simplified to just starting the containers and regenerating
the documentation.

The downside of this is the fact that you'll need the image before you
can create the documentation. This means that either the documentation
must be generated after the image is released or a custom image must be
used that has a pre-release image in it. Since changing images is easier
and significantly faster than building MaxScale from source, this is an
acceptable compromise between ease-of-use and portability.