Added upgraded uncrustify.cfg suitable for uncrustify 0.74
The old config produces the same results as the new config
with the exception of two files: memcachedstorage.cc and
redisstorage.cc, which are improved if formatted.

The config can be compared to the default.cfg in the
uncrustify repo: documentation/htdocs/default.cfg

Limit what directories the script maxscale-uncrustify.sh
touches (which is unecessary because the script will
not be run for mass format and push).