Remove `fibers` and upgrade `sass-loader` to `10.2.1`
`fibers` is not supported on nodejs version > v16.x.

Update sass-loader to 10.2.1 solve it without migrating vue-cli.
https://github.com/webpack-contrib/sass-loader/releases/tag/v10.2.1
Migrating vue-cli to latest version will break HMR (Hot Module Replacement).
For the sake of simplicity, it's better just to update sass-loader.
The migration can be done at a later point where it is stable.