MXS-4043 Filter out `protocol` and `service` params
There are changes from the API since 6.0 that include `service`
in the listener parameters.
The GUI has a separate `service` dropdown and the parameters table
is generated automatically which causes duplicated service input.
The same thing with the `protocol` parameter, but this parameter was
disabled in the GUI. Both parameters should be removed from the
parameters table.