MXS-3990 Explicitly cancel dcalls
Existing dcalls will be cancelled in ~Callable. However, as
at that point the parts of a class derived from Callable will
already have been destructed, the call will just be deleted
without the delayed call first being called with Callable::CANCEL.
That may cause a leak if, when cancelled, the callback deallocates
something. Thus, a warning is logged and thus dcalls should be
explicitly cancelled.