MXS-3722: Add `ddl_only` to binlogrouter
As the binlogrouter already buffers transactions before writing them on
disk, the ddl_only mode can be easily implemented by simply discarding
all non-DDL transactions.

The GTIDs generated in this mode will not increase linearly but the
replication in MariaDB can handle this. The event positions in the
binlogs will be sequential so the generated files end up having no gaps
even if the real binlog positions would create them. This works because
the binlogrouter writes its own positions into the binlog file as it
also generates (or doesn't generate) events in some cases.