Don't measure time in idle_pool_max_conns
The timing values are wildly different for different environments and
they depend largely on the network latency. Given that we can't
guarantee that the latency is below some threshold, it's better to
remove the check for it from the test rather than adjust the limits
upwards to make the test pass. Increasing the limits either makes them
useless or causes false positives when the network is having a bad day.