MXS-4051: Add RouterSession::set_reply
The schemarouter wrote directly to the client DCBs protocol object which
was very much wrong. The other routers called the upstream components
clientReply function before it had the chance to return from the
routeQuery function.

All of these are wrong and the correct thing to do is to queue it up as
the next response using session_set_response. This necessitated the
addition of it for routers as well.

Removed the old service member from the struct as well, it wasn't used
for anything.