MXS-3844 Add `padding` attribute to header of `data-table`
The `/dashboard/servers` view shows a lot of columns,
each column has 24px as padding left and padding right
which is unnecessary. To have more space to render
the text, each column's padding should be customizable.

This adds `padding` attribute to the header object allowing
to pass custom css padding.