MXS-4051: Collect only current resultset rows
This both reduces the overall memory usage as well as allows a read-only
view of the data to be used. The two users of this feature,
readwritesplit and schemarouter, both used it in a way that was easy to
adapt to the new behavior.

The read-only view part of this is done in a separate commit. Doing it
requires changes in both readwritesplit and schemarouter as they use
const references to std::strings.