MXS-4051: Use string_view in schemarouter
The relevant parts of the resultset handling and shard lookup code now
use string_views.

Doing the tolower transformation during the insertion is much more
efficient but it does cause a change in how the two custom commands,
`SHOW DATABASES` and `SHOW SHARDS`, present their output. Given that one
should not rely on case-sensitivity of the names in the first place,
it's not likely to cause problems.