MXS-3960 Prevent anchor link from being clicked while dragging
If the user places the mouse at the anchor link in the node and starts
to drag and drop, the click event will propagate to the anchor link which
is not expected.

This passes `isDragging` state down to `rect-node-content` props so it
can be used to conditionally add `no-pointerEvent` css helper class to
disable pointer event while dragging.