MXS-3844 Fix the padding of the filter index column and editable cell padding
The previous commit that adds `padding` attribute to the header of
`data-table` broke the padding set to the filters table in `relationship-table`
component. The `padding` attribute was actually added long time ago
but it wasn't written well. The previous commit fixed that.

So, now in the `relationship-table`, the padding for the index column doesn't
need to set `!important` rule anymore. In addition, the padding for
editable cell should also be handled in `cellLevelPadding`