MXS-4053 Ignore multi-statements
Multi-statements are always sent to the backend and their result
is not cached. In principle it would be possible to cache the
the result of e.g. a vanilla multi-statement like "SELECT a FROM t;
SELECT b FROM t", but it's not entirelly trivial to be absolutely
certain that is the case, so for now we just ignore them.