Remove global qc_sql_mode value
The variable was written to by multiple threads at the same time and
could cause the wrong value to be returned if two different sql_mode
values were ever used. In addition, both the sql_mode functions should
only work on the thread-local value and the getter function always
returned the global mode.

Since the value is given during the query classifier construction, it
can always be asked from the classifier itself.