Always persist the type parameter for monitors
This was broken by the merge from 6 that contained the fix for
MXS-3962. Now that only modified parameters are saved, the `type`
parameter must be explicitly listed in the values that are always
stored. The reason why it broke was that monitors in 6 did not use the
new mechanism so it was not automatically added.