Adjust the states icons
This uses the `maintenance` icon for the "maintenance" state of the server.
This also uses the `stopped` icon for representing the `stopped` state of
the monitor and service.
The size of the state icons in the tables are increased to 16px to match
with MariaDB design system.

This uses the monitoring application as a reference.