MXS-3936 Only expect as many status replies as requests
It is not certain this is the cause for MXS-3936 as I could not
repeat the error, but expecting more replies than there were
requests sent will cause exactly the reported error. As the code
did not send a requests in all circumstances (even though I am
not sure just how that could be triggered), then it is prudent
to take that into account when waiting for responses.