MXS-3961 Add `draggableGroup` props
When `auto_rejoin=false`, rejoinable servers should be draggable
to the `tree-graph`.

This adds the mentioned props to allow dragging from another list
to the `tree-graph`.

This also renames some drag events and its handler to make them clear
and easier to reuse when the upcoming joinable-servers component is added.