MXS-3961 Replace `swapNodeHeight` with `triggerRerenderNodes`
Swapping nodes' heights works for the switchover action only when
both nodes' heights aren't changed before the action. i.e. If one
of the nodes has the extra info section collapsed, this won't work.

This uses an `uniqueKey` for each `server-node` to trigger a
re-render after the order of the nodes has changed.