MXS-4067: Use proxy protocol for internal connections
This makes MaxScale more conformant to the specification, even if the
use of the UNKNOWN protocol is discouraged. Given that we cannot know
the IP or the port before the connection is created, this is the best
solution we can use.

MariaDB treats the UNKNOWN protocol as an instruction to ignore it
completely which results in the same behavior as before. For Xpand, this
will be needed as it will most likely reject non-proxied connections
from hosts where proxied connections are accepted from.