MXS-4074 Update bootstrap status at each monitor tick
Since the health check of the volatile servers is made independently
of the actual monitor tick, it's possible, especially if the health-
check round, for whatever reason, takes longer than normally, to
end up in a situation where, whenever the Monitor tick occurs,
we are in the middle of performing a health-check round. The effect
of that problem disappears if we simply update the state of the bootstrap
servers from the volatile servers at each Monitor tick, irrespective
of what the state of the health-check round is at that point.