MXS-4010: Add automatic removal of old data files
Using the event timestamps is similar to how MariaDB purges binary logs
if expire_logs_days is configured. By default files are not
automatically removed as this would cause upgrades from older versions
to break. When enabled, the purge operation takes place whenever an
event arrives for a yet-to-be-opened table. This will take place
whenever a file rotation takes place due to either a manual request or
an automated one due to the file size being exceeded.

Only the table that received the event will be inspected: this is done
to speed the process up when there are multiple tables and to prevent
removal of data for inactive tables. If inactive tables were included in
the purge, it would be possible for them to be removed even if no events
for them have arrived. Whether this would be a desirable feature is not
known at this point.