MXS-4067: Update mxs3342_proxy_protocol_pool
Now that MaxScale uses proxy protocol for internal connections, it
cannot immediately connect to the server as the proxy_protocol_networks
variable is not yet set to the correct value. This caused authentication
to fail as there were no valid servers from which to query users
from. Waiting for one monitor tick after the variable is correctly set
fixes it.

Also made the error messages unique. This way it will be obvious which
part is failing.