MXS-4010: Add file rotation test case
The test checks that both the manual and the automated file rotation
work. It also checks that the file purge works as expected.

The timing of the file purge isn't tested that thoroughly but it checks
that old files are pruned. The test case needs to be extended to also
cover longer periods of time than just one second.