MXS-2347: Add session restarting
The /sessions/:id/restart endpoint causes a session to re-open the
internal connection the next time the connection is idle. The
/sessions/restart endpoint does the same for all open sessions.

This is essentially an internal reconnection done at the service level:
all router and filter sessions will be restarted but the client protocol
state will remain the same. This means that the session command history
will also be saved which makes this a valid way of restarting all
sessions.