MXS-3490 Detect and handle group change explicitly
When a group change is detected, we basically stop doing what
we were doing and wait until we again can speak normally to the
cluster. Now, there will be no warnings/errors in the log, but
merely a note that a group change has been detected followed
by a note when it is over.