MXS-2347: Add MaxCtrl commands for reloading sessions
The commands are named reload to make use of the existing reload command
family. Technically it is a reload and a side-effect of it is that the
session is restarted.

The endpoints and functions could be renamed but leaving them as
`restart` better separates them from the alter commands that modify the
current session without restarting the session internally.