MXS-4059 Don't filter fields for Y-axis
Instead of showing only numeric fields(table column), the user should
be freely able to choose any fields for the Y-axis. The graph
should recognize the data types of both axes and draw the graph
accordingly which will be done in the next commits.

This commit removes the filter and the number sign `#` as it's not
needed.