MXS-3619: Synchronize maintenance mode
The maintenance mode of servers is now synchronized across MaxScale
instances. This assumes that a server being in maintenance mode on one
MaxScale is applicable on other MaxScales as well. Since cooperative
monitoring is the suggested method for dealing with multiple MaxScale
instances, this should guarantee that the operation will always work on
all MaxScales.

If setting a server into maintenance mode fails, it means that it
would've been wrong to do so in the first place. This prevents a broken
MaxScale setup from interrupting other MaxScale instances.