MXS-3619: Synchronize administrative users
Storing the administrative users in the JSON configuration makes it
possible to revert back to the original admin users before configuration
synchronization was enabled. This helps avoid irreversible changes to
the configuration files.

Once admin user management is added to the GUI, it will be possible to
configure a whole cluster of MaxScale from any individual MaxScale. The
exception of course are that any files that are read from disk will
still need to be manually synchronized.