MXS-3619: Synchronize draining state
The drain state is also something that users assign which means it must
also be synchronized. The test code for the two is largely identical as
the operations are very similar.

As the configuration synchronization now relies on the actual string
form of the states, they were moved into the mxs::Target class as static
variables in the hopes that changing one will also allow any users of it
to be changed at the same time.