MXS-3152: Add `maxctrl destroy session`
This makes it possible to kill off sessions without having to provide
database credentials. This also allows the session TTL mechanism to be
used to give the session some time before it is closed.

Added test cases that make sure both the normal and TTL versions of the
command work.