MXS-4063 Pollable must return its file descriptor
A step in the process of replacing Worker::[add|remove]_fd() with
Worker::[add|remove]_pollable(). The reason for that is that we need
to have the pointer to the Pollable available when removing a file
descriptor, in the near future when it is made possible to prematurely
abort the reading of data from a socket.