MXS-4063 Remove public Pollable::owner
Instead provide a method polling_worker() that returns the
worker responsible for polling the Pollable. If the Pollable
is not being polled, then nullptr is returned.

As a consequence, the pointer in Pollable can no longer be
used for determining the owner of the DCB and thus an explicit
m_owner was added to DCB. But it should be evaluated whether
it could be removed, as it is primary used in asserts. When
a DCB is not being polled, the owner should always be
implicitly known.