MXS-4063 Handle interrupted poll event handling
If the return value of the event handler of Pollable has
the bit poll_action::INTERRUPTED set, the Pollable together
with the event bitmask is stored in a data structure.

When epoll_wait() returns the next time, if there are such
pending Pollables they will be called after any new events
for other file descriptors have been handled.

This makes it possible in the case of an edge triggered
non-blocking file descriptor, to stop reading before
the reading fails with EWOULDBLOCK. With this arrangement
it is possible to prevent one client from hogging a
routing worker, even if the client writes so fast that
there always is something read.