MXS-3159 Clean up routes
Currently, the value for the refresh-rate component is stored in memory.
It should be stored in localStorage. To reduce repeated code,
there will be a mixin to store the value based on the route name.
At the moment the routes are defined separately in different files,
to use the constant name for route name, it's better to place all routes
in one file instead of having multiple files so that the constant
name can be declared once.