MXS-3398 Remove special handling of 'session_track_system_variables'
A router (or filter) can now tell the service what variables need
to be tracked. This now works both with static and dynamic
configurations including such where the server is linked to the
service after the service has been created.

Note that this only deals with how the tracking of variables is
enabled, not with how the value of a variable affects something.
That is fine as far as @@global.session_track_system_variables
is concerned as the checking of that is hardwire into RWS.