MXS-4095: Fix cdc.py error detection
The error is always in the first three bytes of the first message
received in the response. This means that by reading a whole line at the
start and only looking at the first three bytes makes sure that we catch
the error when it arrives and we don't misinterpret field names for
errors.