MXS-3912: Add timestamps to users JSON
The creation date, last password update and last login are now stored in
the UserInfo class. The values for creation and last password update are
persisted as a side-effect of the operations that modify them but the
last login isn't.

The timestamps are both displayed and stored as standard HTTP date/time
strings (https://www.rfc-editor.org/rfc/rfc2616#page-20). This makes it
easier to display them without having to convert them on the client side
while still allowing localization due to it being in a standard
format. The need to parse the date/time strings into integer values also
prevents any problems that UNIX timestamps might have but this is not a
practical problem. As the loading only happens once per restart (or once
per sync with config sync enabled), the cost is minimal.