MXS-3912: Update user timestamps
The last_update timestamp is now updated whenever the password is
changed. This information is persisted as it will be saved in the JSON
output. The last_login timestamp is updated once per successful
authentication but it is not persisted as it would mean updating the
on-disk file every time someone logs in.