MXS-4087 Replace `content` slot name with "default" slot for `collapse` component
There will a component called `sessions-table` with `<collapse/>` as a wrapper.
That component will be rendered in the dashboard, server and service detail page.
But in the dashboard page, it doesn't need to render `<collapse` component as a wrapper.

This commit helps to get rid of declaring `<template v-slot:content>` which eases
the process for creating `sessions-table` in the next commit.