Set GTID position in select_master tests
The backend server don't always have the whole history in the binlogs,
which means it is better to start from the current GTID position. This
makes sure the tests work regardless of the order in which they are run:
pinloki_upgrade prunes the binlogs which caused pinloki_passwords to
fail.