MXS-4113: Get dependency names from Param
Using the get_dependencies() function the configuration processing code
can construct the objects in the right order when a parameter depends on
some object. This allows the configuration system to express its
dependencies that were previously hard-coded into the configuration
code.

This mechanism can be used to implement both ParamTargetList and
ParamServerList type parameters (found in `develop`) without having to
explicitly add them to the configuration code. It is also required to
conditionally treat a value as a dependency in case only certain values
require objects to be present: this is the case with namedserverfilter
where the special values `->all`, `->master` and `->slave` do not
require an object.