MXS-4113: Expose namedserverfilter dependencies
The ParamHintTarget parameter declares both servers and services as its
dependencies unless the given value is one of the special constants. As
the hints use the name of the object to tell where it should be routed,
the only functional part that needs to be changed in addition to this is
to use mxs::Target::find in the validation code which finds both servers
and services.