MXS-4041: Add REST API TLS certificate reloading
By using the GnuTLS certificate callback mechanism, the certificate that
is used can be changed without having to restart the REST API. It also
validates the certificates before taking them into use which makes it
possible to safely reload the certificates without discarding the old
ones first.

The cert_callback function is called by libmicrohttpd every time it must
serve a TLS certificate to a client. As the check itself does the swap
of the certificates, there is no concurrent access to the currently used
certificate since the daemon is single-threaded.