MXS-4041: Add test case
The test verifies that the encryption for the REST API and reloading of
the certificates work.

It also contains a small stress test to cover the case when there are
multiple parallel requests while a reload takes place. This helps verify
that any changes done in the libmicrohttpd threading model by us won't
break the reloading by allowing concurrent access to the libmicrohttpd
daemon.