Remove assertions from sescmd routing
If transaction_replay is enabled and the service that the readwritesplit
belongs to is configured with the cache filter, the assertion will fail
if autocommit is disabled and a cached result is used for the first
statement of the transaction.

The assertion can happen with any filter that can return early from
routeQuery. In these cases the assumptions the assertion makes are not
true and extending it to cover these cases makes it next to
useless. This means the best action is to simply remove it and to solve
the underlying problem.