MXS-4117 Add mutation creator for syncing wke to flat state of a query module
When changing a worksheet, its flat state of query and queryConn modules
need to be updated. The mutation for doing this for both modules follows
a similar pattern, so a mutation creator can be created to DRY code.

This also refactors `syncedStateMutationsCreator` to use the reduce method
to create object.