MXS-4079: Document conflict probabilities
The numbers for two MaxScales are based on the binomial probability of
getting exactly half of the servers. The retry probabilities were also
easily derived from the base 25% probability of monitors choosing the
same wait interval.

The rest were calculated using brute force and were created by a script
that generates the set of all possible lock distributions and then the
set of conflicting lock distribtions and uses them to calculate the
probability of a conflicting set. A similar script was used to generate
the retry probabilities for four or more MaxScales.

As the number of MaxScales increases, the probability for conflicts also
increases, starting with an average 20% probability with two MaxScale,
55% with three, approaching 100% probability as the number of instances
increases. The random wait period after it is what reduces the
likelihood of conflicts even if the initial conflict probability is
high. This makes the system work even with relatively large numbers of
MaxScale instances.