MXS-4117 Make mutation creators for synced states reusable
The mutation creators were made to use for partial modification of
a worksheet object in the persisted state `worksheets_arr`. However,
when the `querySession` module is introduced, there is a need to
sync states from `queryConn, editor, queryResult` to the `querySession`
module. Hence those mutation creators should be made reusable.