MXS-3398 Introduce Dependency
A Dependency (actually some derived class) expresses a dependency
between a particular MaxScale configuration parameter and some
particular server variable.

With a Dependency it can be arranged so that the value of a
MaxScale configuration parameter automatically tracks the value
of a server variable, either directly or as the result of some
computation.

In this commit, the dependencies are not yet processed.