MXS-4025 Make `openBgConn` action and `getBgConn` reusable
This makes them reusable and renames them to `cloneConn, getCloneConn`, respectively.

This also adds `clone_of_conn_id` attribute when cloning a
connection to make it easier to find the accurate background
connection. This is because, in multiple session tabs,
there would be a case where multiple background connections are
created when the user stops the running query in multiple
session tabs. Checking the connection name is not accurate anymore.