MXS-4025 Add `session-toolbar`
This moves the Run/Stop, visualize buttons and the max rows
dropdown from `worksheet-toolbar` to `session-toolbar`.

When the user runs queries in a session tab, then switches
to a new session tab, some getters such as `getIsConnBusy,
getLoadingQueryResult, getIsStoppingQuery` return wrong value
as those getters use the current active session id to get
the value. So those buttons need to be assigned with a session id.