MXS-4025 Handle killing background connection among sessions tabs and worksheets.
Since worksheets are not rendered all at a time, but rather replacing
a worksheet component will new data. The watcher for `isQueryKilled`
should be manually added and removed when the component is activated
and deactivated.

On the other hand, the `session-toolbar` won't be replaced but
rendering all sessions among worksheets at a time so that the
watcher for `isQueryKilled` is still running even when the worksheet
is inactive.