MXS-3398 Add 'auto_tune' parameter for controlling the auto tuning
If 'auto_tune' is 'none' then no auto tuning will be made, if it
is 'all' then all auto tunable parameters will be tuned.

If this gathers popularity and people want more control, then
this is easy to extend to

auto_tune=connection_keepalive, this_param, that_param

This can now only be modified at startup, which feels as the
right thing to do as its value affects how the configuration
mechanism works.