MXS-4025 Add `getSqlConnId` to get sql connection id from axios config object
`$queryHttp` should be used for requests that may cause busy sql connection
and so that the axios interceptors can know which sql connection is querying.
From axios config url, the connection id can be found, with that id,
the session id can be then found and used for updating `is_conn_busy_map`.

The is needed so that queries can be executed parallelly between sessions
and worksheets.