MXS-2662: Remove key and IV from constructors
The mxb::Cipher has the EVP_CIPHER_CTX which is needed for multiple
encrypt/decrypt cycles but the keys in it are used for only one
time. This means that the keys must be passed to the encrypt and decrypt
functions separately.

Moved the code from the header into the source file as the functions are
likely to get larger later on when more encryption modes are added
(e.g. GCM requires a custom class).