MXS-2662: Add constants for cipher mode selection
This way the configuration variables can be easily mapped to the cipher
modes without having to have the key length be a name of the value. The
required key length can be derived from the size of the encryption key.

Also moved the new key creation into the mxb::Cipher class as it will be
needed for key rotation.