MXS-2662: Add GCM mode to mxb::Cipher
This makes it possible to have authenticated encryption of the binlogs
which give stronger guarantees of consistency. Even without
authenticated encryption, the binlogs would be consistent as they
usually have the 4 byte CRC enabled for verification purposes.

The additional authenticated data for the GCM mode is now hard-coded to
be the last 4 bytes of the IV. GCM mode uses a 12 byte IV and as the
binlogs use 16 bytes (position + 12 byte IV), the extra 4 bytes can be
used for AAD. The current implementation adds the needed 4 bytes to the
IV size. Another option would be to ignore the position and only use the
nonce part of the START_ENCRYPTION event. If needed for other encryption
purposes, the interface for providing custom AAD can be made public.

This commit also adds a helper function that calculates the requires
size in bytes for the encrypted data.