MXS-2662: Add encryption unit test
The test makes sure that data is successfully encrypted and decrypted in
all modes. It does not yet check that data corruption is detected in GCM
mode.

Added stringification helpers and a more convenient error message
function for printing encryption errors. The errors are also now logged
on a single line so that errors from multiple concurrent operations do
not get mixed up. Given that there is usually just a single error, the
output should remain identical with older versions.