Fix crash on KILL command
If a delayed call attached to a session releases the last reference to
the session, the DCall object is deleted while it is still in the middle
of a function call. To avoid this, we can use a `lcall` to return the
execution to the epoll loop before returning the reference.