MXS-3398 Change `auto_tune' to take string list
Now `auto_tune` takes `all` that turns on auto tuning of every
auto tunable parameter or a list of auto tunable parameters that
should be automatically tuned.

Currently there is a bit of unnecessary work being done, since
currently the parameters to be auto tuned do not change after
startup and are the same for all services. However, as the tuning
is done quite rarely, it does not matter.