MXS-4138: Fix pinloki race condition
The sending of events now checks whether the session is still alive
before touching any of the code. The use of the std::weak_ptr isn't
ideal but using a MXS_SESSION reference won't work for testing where
this isn't easily available. A better solution for this is in the long
term is to use the `lcall` function in newer code.