MXS-4100 connection_keepalive=0 causes a memory leak
Queries were pushed to the m_track_queue when tracking
was not enabled.

NOTE: The real reason was that something in the code
causes m_reply.state() to be set differently when
tracking is off versus on. When tracking is on simple
queries are never tracked. This may or may not be the
correct behavior for that state...