MXS-2662: Add KeyManager
The KeyManager class manages versioned encryption keys. The keys are
partitioned by a key ID that can be used to give different services
different encryption keys.

The keys can be rotated while still allowing access to the older
versions of the keys. For the binlogrouter, this means that newer
binlogs are encrypted using a newer key while still allowing old binlogs
to be decrypted.

The keys used to encrypt the data are actually generated by MaxScale
itself which are then stored in an encrypted JSON file. The JSON file is
encrypted using the MasterKey class that implements the actual
encryption. This is similar to how the AWS KMS plugin for MariaDB stores
keys as well as how MongoDB deals with encryption.